+++ model/src/main/scala/org/broadinstitute/dsde/rawls/RawlsException.scala
@@ -1,7 +1,6 @@
 package org.broadinstitute.dsde.rawls
 
 import org.broadinstitute.dsde.rawls.model.ErrorReport
-import spray.http.{StatusCodes, StatusCode}
 
 class RawlsException(message: String = null, cause: Throwable = null) extends Exception(message, cause)
