+++ model/src/main/scala/org/broadinstitute/dsde/rawls/model/RawlsEnumeration.scala
@@ -1,0 +1,6 @@
+package org.broadinstitute.dsde.rawls.model
+
+trait RawlsEnumeration[T <: RawlsEnumeration[T]] { self: T =>
+  def toString: String
+  def withName(name:String): T
+}
